[Sources] Use `git pull` with the `--ff-only` flag
This fixes issue where since we were using `--no-commit`, git was
starting a merge but not committing it. Then subsequent `git pull`'s would
result in the following error from git:

> You have not concluded your merge (MERGE_HEAD exists).
> Please, commit your changes before you can merge.

This commit will mean that `--ff-only` is passed to `git pull`  which means
it wont rebase or merge. It will only allow a fast forward to progress
with the pull.

Fixes #2024